Add Jenkins pipelines
This adds multiple pipelines:

- `build` is responsible for building binaries, syntax checks and
running unit tests. When building a tag, it also publishes binary
artifacts to S3.

- `release` makes the actual releases for a given tag. Releasing is
the act of copying S3 tag artifacts into the relevant buckets (eg.
https://downloads.dcos.io/binaries/cli/linux/x86-64/dcos-1.12/dcos)

- `integration-tests-{platform}` runs integration tests for a given
platform (linux, mac or windows). In order to do that it firsts sets up
a cluster in AWS using DC/OS e2e. It also runs unit tests, as
unfortunately they currently require a cluster to be configured in order
to pass.

- `nightly-integration-tests` runs master branch integration tests on a
daily basis and for all platforms.