Updated 'ls' integration tests to only look for 'stderr' and 'stdout'.
Depending on the security mode of the cluster there might also be
'.ssl*' and '.secret*' files in the sandbox directory in addition to the
standard 'stderr' and 'stdout' files. The 'stderr' and 'stdout' files
will always be present in all cases.

Since we are only trying to test to make sure that ls works, it makes
sense to only make sure that some subset of the files that we can
guarantee to always be there will be present.