Cleaned up the logic for how we assert output in 'ls' integration tests.
The number of output lines in the 'ls' tests depends on the number of
tasks we are 'ls-ing' at the same time. The logic in these tests now
reflects that so that if new test apps are added in the future, the code
in these tests does not have to change.