Added `task attach` subcommand.
* Refactored `TaskIO` object to be friendlier for adding `task attach`.

* Added `task attach` subcommand.

* Changed `get_container_id()` to take a task_id instead of a task_obj.

* Added an exit sequence for `dcos task attach`.

The default exit sequence is 'ctrl-p ctrl-q'. This cannot be changed at
the moment. In the future we should consider allowing this to be
overridden on the command line for `dcos task attach`, similar to Docker.

* Force a resize to 0 before actual resize in `dcos task exec/attach`.

We only do this during the initial resize in order to allow programs
(such as bash) to handle the resize in order to redraw their prompt.

Unfortunately, there is a bug in Mesos at the moment that prevents this
from working properly in the 'Command Executor'. However, it does work
for commands launched with the pod (a.k.a. default) executor. Once this
bug is fixed, this should already be set to work appropriately.

The limitation of this not working appropriately at the moment is that
when running `dcos task attach`, you are not able to resize the screen
and have it react properly. Additionally, you have to hit enter once
after attaching in order to see a prompt if you are attaching to a
shell.

* Updated `popen_tty()` to optionally run commands through a shell.

* Updated all `dcos task exec` tests to not run on Windows.

* Added an integration test for `dcos task attach`.