Added a 'publish_index.py' script to publish the artifacts index to S3.
This script should nto be needed very often. It is only necessary if
changes to the 'index.html' or any of its assets (other than
'artifacts.json') change and need to be reuploaded. A subsequent script
will include the logic to generate 'artifacts.json' and upload it so
that the index dynamically discovers what artifacts it should publish
and display them.