Fix unknown subcommand error for `dcos service`
Currently the service subcommand prints the following in case of an
invalid subcommand:

$ dcos service qwerty
Getting the list of services from `dcos service` is deprecated. Please
use `dcos service list`.
Error: unknown command "qwerty" for "list"

This updates it to print the help menu, an error message, and return
with a non-zero exit code as well.

https://jira.mesosphere.com/browse/DCOS_OSS-3500