Fix unknown subcommand errors (#261)
* Fix unknown subcommand error for `dcos node`

Currently the node subcommand behaves like this:

$ dcos node qwerty
Getting the list of nodes from `dcos node` is deprecated. Please use
`dcos node list`.
Error: unknown command "qwerty" for "list"

This commit updates it to return a proper error for unknown subcommands,
as well as a non-zero exit-code:

$ dcos node qwerty
Usage:
  dcos node [flags]
  dcos node [command]

Commands:
  decommission
      Mark an agent as gone
  diagnostics
      Use diagnostics bundles
  dns
      Return the IP address(es) corresponding to a given hostname
  list
      Show all nodes in the cluster
  list-components
      Print a list of available DC/OS components on specified node
  log
      Print logs for the leading master node or agent nodes
  metrics
      Display the metrics of a node
  ssh
      Establish an SSH connection to the master or agent nodes of your
DC/OS cluster

Options:
      --field dcos node   Name of extra field to include in the output
of dcos node. Can be repeated multiple times to add several fields.
  -h, --help              help for node
      --json              Print in json format

Use "dcos node [command] --help" for more information about a command.

Error: unknown command qwerty

https://jira.mesosphere.com/browse/DCOS_OSS-3500

* Fix unknown subcommand error for `dcos job`

Currently the job subcommand prints the help menu and returns with a
zero exit code in case of invalid command.

This updates it to print an error message and return with a non-zero
exit code as well.

https://jira.mesosphere.com/browse/DCOS_OSS-3500

* Fix unknown subcommand error for `dcos service`

Currently the service subcommand prints the following in case of an
invalid subcommand:

$ dcos service qwerty
Getting the list of services from `dcos service` is deprecated. Please
use `dcos service list`.
Error: unknown command "qwerty" for "list"

This updates it to print the help menu, an error message, and return
with a non-zero exit code as well.

https://jira.mesosphere.com/browse/DCOS_OSS-3500

* Fix usage template

This removes the double usage line when a command is both runnable and
has child commands, as this usually happens for backwards compatibility
reasons and is not something that should be reflected on the help menu.

* Fix test fixtures for help menu