Send an empty process IO data on STDIN EOF
For an `ATTACH_CONTAINER_INPUT` call and when STDIN reaches EOF,
Mesos expects an empty process IO data message, as highlighted in:

https://github.com/dcos/dcos-core-cli/blob/7b67c5a369ff92d1c7776fa5f2b1560406a6d27d/python/lib/dcos/dcos/mesos.py#L1657-L1661

This is currently not done in the new attach and exec commands in Go. It
causes them to exit with an erroneous exit code in some cases. It can be
seen by reaching EOF while running a cat command for example. The
following command returns with an exit code of 137 instead of 0:

    dcos task exec -i my-app cat <input.txt

However, sending an empty process IO message turned out not to be doable in JSON,
due to mesos-go omitting the field when it's empty:

    {"type":"PROCESS_IO","process_io":{"type":"DATA","data":{"type":"STDIN"}}}

Mesos then returns:

    Error: malformed request: Expecting 'process_io.data.data' to be present

To get around this issue, we're switching to protobuf.