Increase the buffer size when reading from piped stdin
We had setup a 1 byte buffer for reading from stdin to work around an
issue with the escape proxy reader we're using. However @vespian
reported that it affects performances of the `dcos task exec` command
quite badly when the input is piped and has a significant size.

This updates the code to use a 1 byte buffer only for the terminal
use-case, and increasing it to 512 for piped input.

Running the following command with a ~50MB file showed the following
results:

    dcos task exec -i my-test sh -c "cat >/dev/null" < /path/to/big/file

Former Python command: 0.67s user 0.07s system 2% cpu 32.716 total

With this patch: 0.34s user 0.20s system 2% cpu 24.880 total

The Go command prior to this patch hadn't finished after 2 minutes.

https://jira.mesosphere.com/browse/DCOS_OSS-5357