Fix error when detecting a partial escape sequence
We've been using a 1 byte buffer with the proxy reader to get around a
failure in one of our integration tests.

However this creates an error condition, when the CLI matches a partial
escape sequence, it cannot recover from it and will fail with a stack
trace.

This can be reproduced by taping eg. `ctrl-p+A`, when the proxy reader
reads `ctrl-p`, it will not return any data as it's matching the
beginning of the `ctrl-p+ctrl-q` escape sequence. However, the next
character `A` should lead to the reader returning `ctrl-p` and then `A`,
either on the current read if the buffer has enough capacity, or on the
next read. However this doesn't seem to work.

Instead of using a 1 byte buffer, I've patched the escape proxy reader
to work with a base reader potentially returning more than 1 byte (eg. a
pty), and will submit a PR to Docker.

https://jira.mesosphere.com/browse/DCOS_OSS-5443