Fix Docker perm issue on Linux
When running `make` using Docker (`NO_DOCKER` is not exported), the
`make plugin` command then fails on Linux because of a permission issue.
The binary is owned by root thus the plugin ZIP cannot be created.

This changes the Docker command to run as the current user, it also sets
the HOME directory to `/tmp` as `/root` wouldn't be writable but is
still needed for Go cache files.