Revert "Refactor the rest of `dcos package repo...` into Go (#346)"
This reverts commit 0c721636d8722a31c5182e25c2ccf7806863cedf.

We had to mute a test while refactoring the command, this is still
tracked in https://jira.mesosphere.com/browse/DCOS_OSS-5529 and we
didn't identify yet what is the cause of it. It might be a regression
with our Go code.

As DC/OS 2.0 now only receives fixes, we should continue the Go
refactoring on the 2.1 branch.