Remove DC/OS E2E
Currently we have a mix of DC/OS E2E and DC/OS Launch. We pinned an old
version of DC/OS E2E which doesn't work with DC/OS dev anymore, updating
DC/OS E2E to its latest version would probably fix the issue, this was
tried in #376 without success.

As a simple fix we can remove DC/OS E2E and stick to DC/OS Launch for
now, in the long-term we should move to the universal DC/OS installer.