Wait for marathon to be up after changing leadership
Aftere posting DELETE /v2/leader marathon abdicate leadership which
cause marathon to restart. If there is single master cluster (which is
what we are using for CI) this operation can take couple of seconds. On
my local machine it takes over 20s for Marathon to be operational after
abdicating. This change forece test to wait for Marathon to operational
before marking test as PASSED.

Fixes: https://jira.mesosphere.com/browse/DCOS_OSS-5599
Refs: https://github.com/mesosphere/marathon/issues/3251